requirements.txt
# Core Python Libraries
pandas==1.5.3
numpy==1.24.0
scikit-learn==1.2.0
matplotlib==3.6.3
seaborn==0.12.1

# PyTorch and Transformers
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0

# For Hyperparameter Tuning
optuna==3.2.0

# For Image Handling
Pillow==9.5.0
requests==2.31.0

# For Logging and Debugging
loguru==0.6.0

pytest
